finish implementation of langchain utils
