+++ api/deps.py
@@ -1,6 +1,6 @@
 from fastapi import Request
 
-from hybridrag.document_processor import DocumentProcessor
+from hybridrag.document_processor_ingest import DocumentProcessorIngest
 from hybridrag.scraper import PubMedScraper
 from src.db.db_client import QdrantWrapper
 
@@ -9,8 +9,8 @@ def get_db_client(request: Request) -> QdrantWrapper:
     return request.app.state.db_client  # type: ignore
 
 
-def get_document_processor(request: Request) -> DocumentProcessor:
-    return request.app.state.document_processor  # type: ignore
+def get_document_processor_ingest(request: Request) -> DocumentProcessorIngest:
+    return request.app.state.document_processor_ingest  # type: ignore
 
 
 def get_scraper(request: Request) -> PubMedScraper:
