+++ hybridrag/document_processor.py
@@ -1,77 +1,82 @@
+import os
 from typing import List
+from langchain_community.document_loaders import PyPDFLoader
+from langchain_text_splitters import RecursiveCharacterTextSplitter
 from sentence_transformers import SentenceTransformer
 import logging
+
+from src.db.db_client import QdrantWrapper
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
 class DocumentProcessor:
-    def __init__(self, model_name="all-MiniLM-L6-v2"):
+    def __init__(
+        self,
+        db_client: QdrantWrapper,
+        model_name: str = "all-MiniLM-L6-v2",
+        directory_path: str = "",
+    ):
         try:
             self.model = SentenceTransformer(model_name)
+            self.directory_path = directory_path
+            self.db_client = db_client
         except Exception as e:
             logger.error(f"Error initializing SentenceTransformer: {e}")
             raise
 
-    def chunk_text(self, text, max_chunk_size=1000) -> List[str]:
+    def procces_directory(self):
+        file_paths = os.listdir(self.directory_path)
+        inserted_count = 0
+        for file_path in file_paths:
+            pdf_path = os.path.join(self.directory_path, file_path)
+            if pdf_path.lower().endswith(".pdf"):
+                self.process_pdf(pdf_path)
+                inserted_count += 1
+        return inserted_count
+
+    def process_pdf(self, pdf_path):
         try:
-            words = text.split()
-            chunks = []
-            current_chunk = []
-            current_size = 0
-
-            for word in words:
-                if current_size + len(word) > max_chunk_size and current_chunk:
-                    chunks.append(" ".join(current_chunk))
-                    current_chunk = []
-                    current_size = 0
-                current_chunk.append(word)
-                current_size += len(word) + 1  # +1 for space
-
-            if current_chunk:
-                chunks.append(" ".join(current_chunk))
-
-            return chunks
-        except Exception as e:
-            logger.error(f"Error chunking text: {e}")
-            return []
-
-    def embed_text(self, text):
-        try:
-            return self.model.encode(text)
-        except Exception as e:
-            logger.error(f"Error embedding text: {e}")
-            return None
-
-    def process_paper(self, paper):
-        try:
-            full_text = f"{paper['title']} {paper['abstract']}"
-            if not full_text.strip():
-                logger.warning(f"Empty text for paper {paper['pmid']}")
-                return None
+            loader = PyPDFLoader(pdf_path)
+            documents = loader.load()
+            full_text = "".join([doc.page_content for doc in documents])
             chunks = self.chunk_text(full_text)
-            if not chunks:
-                logger.warning(f"No chunks generated for paper {paper['pmid']}")
-                return None
             embeddings = [self.embed_text(chunk) for chunk in chunks]
             valid_chunks_and_embeddings = [
                 (chunk, emb)
                 for chunk, emb in zip(chunks, embeddings)
                 if emb is not None and len(emb) > 0
             ]
-
-            if not valid_chunks_and_embeddings:
-                logger.warning(
-                    f"No valid embeddings generated for paper {paper['pmid']}"
+            pdf_filename = os.path.basename(pdf_path)
+            if valid_chunks_and_embeddings:
+                self.db_client.insert_paper(
+                    pdf_filename,
+                    [chunk for chunk, _ in valid_chunks_and_embeddings],
+                    [emb for _, emb in valid_chunks_and_embeddings],
                 )
-                return None
-
-            return {
-                "pmid": paper["pmid"],
-                "chunks": [chunk for chunk, _ in valid_chunks_and_embeddings],
-                "embeddings": [emb for _, emb in valid_chunks_and_embeddings],
-            }
+                logger.info(
+                    f"Processed and inserted PDF {pdf_filename} into the database."
+                )
+            else:
+                logger.warning(f"No valid chunks or embeddings for PDF {pdf_filename}.")
         except Exception as e:
-            logger.error(f"Error processing paper: {e}")
+            logger.error(f"Error processing PDF {pdf_path}: {e}")
+
+    def chunk_text(self, text, max_chunk_size=500) -> List[str]:
+        try:
+            text_splitter = RecursiveCharacterTextSplitter(
+                chunk_size=max_chunk_size, chunk_overlap=200
+            )
+            chunks = text_splitter.split_text(text)
+            return chunks
+        except Exception as e:
+            logger.error(f"Error chunking text: {e}")
+            return []
+
+    def embed_text(self, text: str):
+        try:
+            return self.model.encode(text)
+        except Exception as e:
+            logger.error(f"Error embedding text: {e}")
             return None
