+++ testing/test.py
@@ -21,12 +21,15 @@ def test_ingest():
     print("\nWaiting for 5 seconds for ingestion to complete...\n")
     time.sleep(5)
 
+
 def test_insert():
     insert_url = f"{BASE_URL}/insert"
-    insert_query = {"directory_path": "src\graph\samplepdfs"}
+    directory_path = r"src/graph/samplepdfs"
+
+    params = {"directory_path": directory_path}
 
     print("Testing /Insert")
-    response = requests.post(insert_url, json=insert_query)
+    response = requests.post(insert_url, params=params)
 
     if response.status_code == 200:
         print("Insert successful:")
@@ -60,5 +63,6 @@ def test_query():
 
 
 if __name__ == "__main__":
-    # test_ingest()
-    test_query()
+    test_ingest()
+    # test_query()
+    # test_insert()
